userdelete

+++ backend/app/Http/Controllers/RoleController.php
@@ -6,7 +6,6 @@ use Illuminate\Http\Request;
 use App\Repositories\RoleRepositoryInterface;
 use App\Models\User;
 
-
 class RoleController extends Controller
 {
     protected $roleRepository;
@@ -18,9 +17,8 @@ class RoleController extends Controller
 
     public function index()
     {
-        if (!auth()->check() || !auth()->user()->can('AccessAdmin', User::class)) {
-            abort(403, 'Unauthorized action.');
-        }
+        // Sử dụng policy để kiểm tra quyền truy cập
+        $this->authorize('accessAdmin', User::class);
 
         $roles = $this->roleRepository->all();
         return view('roles.index', compact('roles'));
@@ -28,34 +26,43 @@ class RoleController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
+        $this->authorize('accessAdmin', User::class);
+
+        $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:roles,name',
         ]);
 
-        $this->roleRepository->create($request->all());
+        $this->roleRepository->create($validatedData);
 
-        return redirect()->route('roles.index')->with('success', 'Role created successfully.');
+        return redirect()->route('roles.index')
+            ->with('success', 'Role created successfully.');
     }
 
     public function update(Request $request, $id)
     {
-        $request->validate([
+        $this->authorize('accessAdmin', User::class);
+
+        $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:roles,name,' . $id,
         ]);
 
-        $this->roleRepository->update($id, $request->all());
+        $this->roleRepository->update($id, $validatedData);
 
-        return redirect()->route('roles.index')->with('success', 'Role updated successfully.');
+        return redirect()->route('roles.index')
+            ->with('success', 'Role updated successfully.');
     }
 
     public function destroy($id)
     {
+        $this->authorize('accessAdmin', User::class);
+
         $deleted = $this->roleRepository->delete($id);
 
-        if ($deleted) {
-            return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
-        } else {
-            return redirect()->route('roles.index')->with('error', 'Role cannot be deleted because it is associated with users or hotels.');
-        }
+        $message = $deleted
+            ? 'Role deleted successfully.'
+            : 'Role cannot be deleted because it is associated with users or hotels.';
+        $status = $deleted ? 'success' : 'error';
+
+        return redirect()->route('roles.index')->with($status, $message);
     }
 }
